Improvements, moved to separate file, improved TableView __init__
